requirements.txt
wagtail==1.3.1
Django==1.8.7
django-fontawesome==0.2.4
django-pyodbc-azure==1.8.7.2
pyodbc==3.0.10
django-multiselectfield==0.1.3
Pygments==2.1
boto==2.39.0
django-storages==1.4
django-htmlmin==0.9.0
